requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchtext>=0.13.0
underthesea>=1.3.4
spacy>=3.4.0
pillow>=9.0.0
pandas>=1.4.0
numpy>=1.21.0
tqdm>=4.64.0
